+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - The Redstone Tool would show an indicator on the linked block twice
 - When a Redstone Interface changes what it's linked to, it now updates both the old and new linked blocks
 - The Redstone Observer and Basic Redstone Interface now properly update on the client when their linked position changes
+- When a Filtered Platform's inventory changes, it's instantly synced to the client, which prevents item entity desync
 
 # 1.4.0